week 4 exercises 4.19-4.21

+++ Osa 4/bloglist-backend/controllers/blogs.js
@@ -1,4 +1,5 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
@@ -11,36 +12,62 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  if (!(request.body.title && request.body.url)) {
-    return response.status(400).json({ error: 'title or url missing' })
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!(request.token && decodedToken.id)) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    if (!(request.body.title && request.body.url)) {
+      return response.status(400).json({ error: 'title or url missing' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const newBlog = {
+      title: request.body.title,
+      author: request.body.author,
+      url: request.body.url,
+      likes: request.body.likes || 0,
+      user: user._id
+    }
+
+    const addedBlog = await new Blog(newBlog).save()
+    user.blogs = user.blogs.concat(addedBlog._id)
+    await user.save()
+
+    response.status(201).json(Blog.format(addedBlog))
+  } catch (exception) {
+    console.log(exception)
+    if (exception.name === 'JsonWebTokenError') {
+      return response.status(401).json({ error: exception.message })
+    }
+    response.status(500).json({ error: 'something went wrong...' })
   }
-
-  const users = await User.find({})
-  const user = users[0]
-
-  const newBlog = {
-    title: request.body.title,
-    author: request.body.author,
-    url: request.body.url,
-    likes: request.body.likes || 0,
-    user: users[0]._id
-  }
-
-  const addedBlog = await new Blog(newBlog).save()
-  user.blogs = user.blogs.concat(addedBlog._id)
-  await user.save()
-
-  response.status(201).json(Blog.format(addedBlog))
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id)
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+    if (!(req.token && decodedToken.id)) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(req.params.id)
+    if (blog.user.toString() === decodedToken.id){
+      await Blog.findByIdAndRemove(req.params.id)
+    } else {
+      return res.status(403).json({ error: 'user permissions insufficient' })
+    }
 
     res.status(204).end()
   } catch (exception) {
     console.log(exception)
-
+    if (exception.name === 'JsonWebTokenError') {
+      return res.status(401).json({ error: exception.message })
+    }
     res.status(400).send({ error: 'malformatted id' })
   }
 })
